osint_suite_pro/requirements.txt
streamlit>=1.20
requests
beautifulsoup4
pandas
fpdf
plotly
openpyxl
networkx
python-dotenv